feat: Добавлена проверка на null в методе findObjectById в классе UserDbStorage, FilmDbStorage GenreDbStorage и MpaDbStorage. В классе FilmDbStorage добавлен batchUpdate в saveGenresToFilm, исправлен sql запрос в getAllObjects. Добавлены тесты : GenreControllerTest и MpaControllerTest.

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/MpaController.java
@@ -26,7 +26,7 @@ public class MpaController {
     }
 
     @GetMapping("/{id}")
-    public Mpa getMpaById(@PathVariable int id) {
+    public Mpa getMpaById(@PathVariable long id) {
         log.info(LogMessages.GET_MPA_BY_ID_REQUEST.toString(), id);
         return service.getMpaById(id);
     }
